refactor: reorganize deployment service imports and structure
